refactor: correct InitailContext naming and WebSettings type
