+++ context/InitialContext.tsx
@@ -1,21 +1,19 @@
 'use client'
-import { getWebSettings } from '@/services/webSettings'
-import { webSettingsData } from '@/types/common'
-import { ReactNode, createContext, useContext, useEffect } from 'react'
+
+import { PropsWithChildren, createContext, useContext, useEffect } from 'react'
 import { useImmer } from 'use-immer'
 
+import { getWebSettings } from '@/services/webSettings'
+import { WebSettingsData } from '@/types/common'
+
 type WebSettingsContextType = {
-  webSettingsData: webSettingsData | null
+  webSettingsData: WebSettingsData | null
 }
 
 const WebSettingsContext = createContext<WebSettingsContextType | null>(null)
 
-type WebSettingsProviderProps = {
-  children: ReactNode
-}
-
-export const WebSettingsProvider = ({ children }: WebSettingsProviderProps) => {
-  const [data, setData] = useImmer<webSettingsData | null>(null)
+export const WebSettingsProvider = ({ children }: PropsWithChildren) => {
+  const [data, setData] = useImmer<WebSettingsData | null>(null)
 
   useEffect(() => {
     getWebSettings().then((res) => setData(res.data))
